Moved `unregisterPluginCustomPostTypes` up
